+++ src/types.rs
@@ -1,5 +1,7 @@
+use chrono::serde::ts_seconds_option::deserialize as from_tsopt;
 use pgrx::prelude::*;
 use serde::{Deserialize, Serialize};
+use sqlx::types::chrono::Utc;
 use sqlx::FromRow;
 use std::fmt::{Display, Formatter};
 use std::str::FromStr;
@@ -102,4 +104,26 @@ pub struct JobParams {
 
 fn default_schedule() -> String {
     "realtime".to_string()
+}
+
+// schema for all messages that hit pgmq
+#[derive(Clone, Deserialize, Debug, Serialize)]
+pub struct JobMessage {
+    pub job_name: String,
+    pub job_meta: VectorizeMeta,
+    pub inputs: Vec<crate::transformers::types::Inputs>,
+}
+
+// schema for every job
+// also schema for the vectorize.vectorize_meta table
+#[derive(Clone, Debug, Deserialize, FromRow, Serialize)]
+pub struct VectorizeMeta {
+    pub job_id: i64,
+    pub name: String,
+    pub job_type: JobType,
+    pub transformer: String,
+    pub search_alg: SimilarityAlg,
+    pub params: serde_json::Value,
+    #[serde(deserialize_with = "from_tsopt")]
+    pub last_completion: Option<chrono::DateTime<Utc>>,
 }
